+++ scripts/macros/2024/classFeatures/bard/glamour/mantleOfInspiration.js
@@ -26,7 +26,7 @@ async function added({trigger: {entity: item}}) {
     if (item.actor.system.scale[classIdentifier]?.[scaleIdentifier]) return;
     if (item.actor.system.scale[classIdentifier]?.['inspiration']) {
         await itemUtils.setConfig(item, 'scaleIdentifier', 'inspiration');
-        await genericUtils.update(item, 'system.activities.healManOfInclass.healing.bonus', 'scale.' + classIdentifier + '.inspiration.die * 2');
+        await genericUtils.update(item, {'system.activities.healManOfInclass.healing.bonus': 'scale.' + classIdentifier + '.inspiration.die * 2'});
     }
     await itemUtils.fixScales(item);
 }