+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ## Update Notes:
 - This is not a V14 update. An update for V14 is still a work in progress.
 ## New Modern Content:
-- Studdied Attack
+- Studied Attack
 - Summer Winds
 ## Bug Fixes:
 - Emboldening Bond